+++ pokekotlin/src/test/kotlin/me/sargunvohra/lib/pokekotlin/test/BulkTest.kt
@@ -3,16 +3,13 @@ package me.sargunvohra.lib.pokekotlin.test
 import me.sargunvohra.lib.pokekotlin.model.ApiResourceList
 import me.sargunvohra.lib.pokekotlin.model.NamedApiResourceList
 import me.sargunvohra.lib.pokekotlin.test.util.mockClient
-import org.testng.annotations.Test
+import org.junit.Ignore
+import org.junit.Test
 import java.util.*
 import kotlin.reflect.memberProperties
 import kotlin.test.fail
 
 class BulkTest {
-
-    companion object {
-        private const val enabled = false
-    }
 
     private fun checkNulls(obj: Any) {
         if (obj is Collection<*>) {
@@ -71,242 +68,290 @@ class BulkTest {
         runTest(list[0].category, list.map { it.id }, getObject)
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkBerry() {
         runTest1({ o, l -> mockClient.getBerryList(o, l) }, { i -> mockClient.getBerry(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkBerryFirmness() {
         runTest1({ o, l -> mockClient.getBerryFirmnessList(o, l) }, { i -> mockClient.getBerryFirmness(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkBerryFlavor() {
         runTest1({ o, l -> mockClient.getBerryFlavorList(o, l) }, { i -> mockClient.getBerryFlavor(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkContestType() {
         runTest1({ o, l -> mockClient.getContestTypeList(o, l) }, { i -> mockClient.getContestType(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkContestEffect() {
         runTest2({ o, l -> mockClient.getContestEffectList(o, l) }, { i -> mockClient.getContestEffect(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkSuperContestEffect() {
         runTest2({ o, l -> mockClient.getSuperContestEffectList(o, l) }, { i -> mockClient.getSuperContestEffect(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkEncounterMethod() {
         runTest1({ o, l -> mockClient.getEncounterMethodList(o, l) }, { i -> mockClient.getEncounterMethod(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkEncounterCondition() {
         runTest1({ o, l -> mockClient.getEncounterConditionList(o, l) }, { i -> mockClient.getEncounterCondition(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkEncounterConditionValue() {
         runTest1({ o, l -> mockClient.getEncounterConditionValueList(o, l) }, { i -> mockClient.getEncounterConditionValue(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkEvolutionChain() {
         runTest2({ o, l -> mockClient.getEvolutionChainList(o, l) }, { i -> mockClient.getEvolutionChain(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkEvolutionTrigger() {
         runTest1({ o, l -> mockClient.getEvolutionTriggerList(o, l) }, { i -> mockClient.getEvolutionTrigger(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkGeneration() {
         runTest1({ o, l -> mockClient.getGenerationList(o, l) }, { i -> mockClient.getGeneration(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkPokedex() {
         runTest1({ o, l -> mockClient.getPokedexList(o, l) }, { i -> mockClient.getPokedex(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkVersion() {
         runTest1({ o, l -> mockClient.getVersionList(o, l) }, { i -> mockClient.getVersion(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkVersionGroup() {
         runTest1({ o, l -> mockClient.getVersionGroupList(o, l) }, { i -> mockClient.getVersionGroup(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkItem() {
         runTest1({ o, l -> mockClient.getItemList(o, l) }, { i -> mockClient.getItem(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkItemAttribute() {
         runTest1({ o, l -> mockClient.getItemAttributeList(o, l) }, { i -> mockClient.getItemAttribute(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkItemCategory() {
         runTest1({ o, l -> mockClient.getItemCategoryList(o, l) }, { i -> mockClient.getItemCategory(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkItemFlingEffect() {
         runTest1({ o, l -> mockClient.getItemFlingEffectList(o, l) }, { i -> mockClient.getItemFlingEffect(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkItemPocket() {
         runTest1({ o, l -> mockClient.getItemPocketList(o, l) }, { i -> mockClient.getItemPocket(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkMove() {
         runTest1({ o, l -> mockClient.getMoveList(o, l) }, { i -> mockClient.getMove(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkMoveAilment() {
         runTest1({ o, l -> mockClient.getMoveAilmentList(o, l) }, { i -> mockClient.getMoveAilment(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkMoveBattleStyle() {
         runTest1({ o, l -> mockClient.getMoveBattleStyleList(o, l) }, { i -> mockClient.getMoveBattleStyle(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkMoveCategory() {
         runTest1({ o, l -> mockClient.getMoveCategoryList(o, l) }, { i -> mockClient.getMoveCategory(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkMoveDamageClass() {
         runTest1({ o, l -> mockClient.getMoveDamageClassList(o, l) }, { i -> mockClient.getMoveDamageClass(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkMoveLearnMethod() {
         runTest1({ o, l -> mockClient.getMoveLearnMethodList(o, l) }, { i -> mockClient.getMoveLearnMethod(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkMoveTarget() {
         runTest1({ o, l -> mockClient.getMoveTargetList(o, l) }, { i -> mockClient.getMoveTarget(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkLocation() {
         runTest1({ o, l -> mockClient.getLocationList(o, l) }, { i -> mockClient.getLocation(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkLocationArea() {
         runTest1({ o, l -> mockClient.getLocationAreaList(o, l) }, { i -> mockClient.getLocationArea(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkPalParkArea() {
         runTest1({ o, l -> mockClient.getPalParkAreaList(o, l) }, { i -> mockClient.getPalParkArea(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkRegion() {
         runTest1({ o, l -> mockClient.getRegionList(o, l) }, { i -> mockClient.getRegion(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkAbility() {
         runTest1({ o, l -> mockClient.getAbilityList(o, l) }, { i -> mockClient.getAbility(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkCharacteristic() {
         runTest2({ o, l -> mockClient.getCharacteristicList(o, l) }, { i -> mockClient.getCharacteristic(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkEggGroup() {
         runTest1({ o, l -> mockClient.getEggGroupList(o, l) }, { i -> mockClient.getEggGroup(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkGender() {
         runTest1({ o, l -> mockClient.getGenderList(o, l) }, { i -> mockClient.getGender(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkGrowthRate() {
         runTest1({ o, l -> mockClient.getGrowthRateList(o, l) }, { i -> mockClient.getGrowthRate(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkNature() {
         runTest1({ o, l -> mockClient.getNatureList(o, l) }, { i -> mockClient.getNature(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkPokeathlonStat() {
         runTest1({ o, l -> mockClient.getPokeathlonStatList(o, l) }, { i -> mockClient.getPokeathlonStat(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkPokemon() {
         runTest1({ o, l -> mockClient.getPokemonList(o, l) }, { i -> mockClient.getPokemon(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkPokemonEncounters() {
         runTest1({ o, l -> mockClient.getPokemonList(o, l) }, { i -> mockClient.getPokemonEncounterList(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkPokemonColor() {
         runTest1({ o, l -> mockClient.getPokemonColorList(o, l) }, { i -> mockClient.getPokemonColor(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkPokemonForm() {
         runTest1({ o, l -> mockClient.getPokemonFormList(o, l) }, { i -> mockClient.getPokemonForm(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkPokemonHabitat() {
         runTest1({ o, l -> mockClient.getPokemonHabitatList(o, l) }, { i -> mockClient.getPokemonHabitat(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkPokemonShape() {
         runTest1({ o, l -> mockClient.getPokemonShapeList(o, l) }, { i -> mockClient.getPokemonShape(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkPokemonSpecies() {
         runTest1({ o, l -> mockClient.getPokemonSpeciesList(o, l) }, { i -> mockClient.getPokemonSpecies(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkStat() {
         runTest1({ o, l -> mockClient.getStatList(o, l) }, { i -> mockClient.getStat(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkType() {
         runTest1({ o, l -> mockClient.getTypeList(o, l) }, { i -> mockClient.getType(i) })
     }
 
-    @Test(enabled = enabled)
+    @Ignore
+    @Test
     fun bulkLanguage() {
         runTest1({ o, l -> mockClient.getLanguageList(o, l) }, { i -> mockClient.getLanguage(i) })
     }