+++ pokekotlin/src/test/kotlin/me/sargunvohra/lib/pokekotlin/test/model/BerryTest.kt
@@ -4,7 +4,7 @@ import me.sargunvohra.lib.pokekotlin.model.FlavorBerryMap
 import me.sargunvohra.lib.pokekotlin.model.Name
 import me.sargunvohra.lib.pokekotlin.model.NamedApiResource
 import me.sargunvohra.lib.pokekotlin.test.util.mockClient
-import org.testng.annotations.Test
+import org.junit.Test
 import kotlin.test.assertEquals
 
 class BerryTest {