+++ pokekotlin/src/test/kotlin/me/sargunvohra/lib/pokekotlin/test/model/ContestTest.kt
@@ -5,7 +5,7 @@ import me.sargunvohra.lib.pokekotlin.model.Effect
 import me.sargunvohra.lib.pokekotlin.model.FlavorText
 import me.sargunvohra.lib.pokekotlin.model.NamedApiResource
 import me.sargunvohra.lib.pokekotlin.test.util.mockClient
-import org.testng.annotations.Test
+import org.junit.Test
 import kotlin.test.assertEquals
 
 class ContestTest {